First stab at implementing controller arbitration
We switched from an asynchronous server implementation to a synchronous
one for the P4Runtime bidi stream, as it makes the code easier to reason
about and there isn't a requirement to support a huge number of
controller connections per switch.

We started implementing controller arbitration, which allows the switch
to maintain multiple controller connections simultaneously (to a single
master controller and several slaves). As of now, some important pieces
of the code are commented out, in order to give a chance to clients to
update their code. When we uncomment this code, we will also add some
unit tests to test the master arbitration state machine.